requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
langchain
langchain-google-genai
chromadb
pydantic
requests
streamlit
langchain-community
